%
%  Copyright (C) 2011-2015 Christina Yassouridis
% 
%

\name{funcyOctave-package}
\alias{funcyOctave-package}
\alias{funcyOctave}
\docType{package}
\title{
\packageTitle{funcyOctave}
}
\description{
\packageDescription{funcyOctave}
}
\details{
funcyOctave is a supplementary package to the R-package funcy to provide
an additional regularization technique for sparse datasets. It is an
interface to the Matlab package PACE, usable only on Unix-platforms. funcyOctave
comes into play if the method-parameter in the function regFuncy (of
package funcy) is set to "pace".
  }

\author{
\packageAuthor{funcyOctave}

Maintainer: \packageMaintainer{funcyOctave}
}
\references{
  http://www.stat.ucdavis.edu/PACE/
}
\keyword{Functional Principal Component Analysis}

