%
%  Copyright (C) 2011-2015 Christina Yassouridis
% 
%

\name{simDat}
\docType{data}
\alias{simDat}
\title{Example data set for irregular data}
\description{
  A simulated dataset containing 100 observations, each evaluated on 4 to 10 time
  points. 
}
\usage{data("simDat")}
\format{Data with curve-ID in first column, curve evaluations in
  second and time points in third column.}

\keyword{data}
